2022_Dataset_convert_CSV_to_YOLOv5_Annotations
This script Reads through the CSV-Guava-Pollinators-percentages.csv file and gets the image name, class and axis for the objects, looks inside the images and using the information creates yolov5 text format annotations for the existing images. Also creates text files which have names of the images those need to be augmented and the name of the images that are to be split in 50:50 ratio (keeping 50% in train set and moving other half in the test set).