Text file containing information about images and labels that needs to be augmented and shifted to test set
both files are created for 2021 and 2022 dataset using convert_json_annotations(2021).py and 2_convert_csv_yolov5(2022).py scripts respectively which are later merged to form to form a single images_to_augment.txt and images_to_shift.txt file for hybrid_dataset.

1. images_to_augment.txt : This file contains detains about the labels of the species that need to be augmented and their image file names which have those species as object in it.

2. images_to_shift.txt : This file contains detains about the labels of the species that are in abundance (overexpressed) and their image file names which have those species as object in it.

these files are used by 4_Move_and_augment(merged_data).py file to look into the training set for the images mentioned in the text file. If the file exists in images_to_shift.txt it moves 50 percent of those files from the training set to test set to remove bias.

On the other hand, if the image name is in images_to_augment.txt, the script augments the respective images in the train set and creates annotations for the augmented images using the main annotation file as reference.